+++ Assets/Mirror/Examples/Tanks/Prefabs/Tank.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 114118589361100106}
   - component: {fileID: 2240606817507776182}
   - component: {fileID: 114250499875391520}
+  - component: {fileID: 7257542793140390888}
   - component: {fileID: 6900008319038825817}
   - component: {fileID: 5194388907919410155}
   - component: {fileID: 114654712548978148}
@@ -52,7 +53,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   serverOnly: 0
-  m_AssetId: 
+  m_AssetId: 6f43bf5488a7443d19ab2a83c6b91f35
   m_SceneId: 0
 --- !u!95 &2240606817507776182
 Animator:
@@ -89,6 +90,20 @@ MonoBehaviour:
   syncInterval: 0.1
   compressRotation: 1
   clientAuthority: 1
+--- !u!114 &7257542793140390888
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1916082411674582}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4b6be33745b1c49caa8b5534fc7d7172, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  syncMode: 0
+  syncInterval: 0.1
 --- !u!195 &6900008319038825817
 NavMeshAgent:
   m_ObjectHideFlags: 0
